WIP separate non-gradio functions
