+++ src/version.py
@@ -1,1 +1,1 @@
-__version__ = "b26a2f0f61240c7685eb94e44fb4a6c9c3965e15"
+__version__ = "1443fcb51d39a3cc13224381617395c767d195c7"
